requirements.txt
azure-servicebus>=7.12.1
pyperclip>=1.9.0
